feat(http): make the request immutable
Make the request immutable to create a new request once we want to modify it.

BREAKING CHANGE: Changed the HttpExecution to specify a generic for the content. Make it immutable,
you need to clone it to modify it. Added the content property that will be now returned by the
executor. The content property is filled by the handler middleware based on the return type of the
request handler. Now, it should return the content instead of the remote response.

fix #59